LayoutTests/platform/glib/fast/canvas/fallback-content-expected.txt

Link  Button
Focusable
This test makes sure that focusable elements in canvas fallback content are focusable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


link1 should be focusable.
PASS document.activeElement == element is true

button1 should be focusable.
PASS document.activeElement == element is true

text1 should be focusable.
PASS document.activeElement == element is true

checkbox1 should be focusable.
PASS document.activeElement == element is true

radio1 should be focusable.
PASS document.activeElement == element is true

slider1 should be focusable.
PASS document.activeElement == element is true

submit1 should be focusable.
PASS document.activeElement == element is true

combobox1 should be focusable.
PASS document.activeElement == element is true

listbox1 should be focusable.
PASS document.activeElement == element is true

textarea1 should be focusable.
PASS document.activeElement == element is true

focusable1 should be focusable.
PASS document.activeElement == element is true

link2 should be focusable.
PASS document.activeElement == element is true

button2 should be focusable.
PASS document.activeElement == element is true

text2 should be focusable.
PASS document.activeElement == element is true

checkbox2 should be focusable.
PASS document.activeElement == element is true

radio2 should be focusable.
PASS document.activeElement == element is true

slider2 should be focusable.
PASS document.activeElement == element is true

submit2 should be focusable.
PASS document.activeElement == element is true

combobox2 should be focusable.
PASS document.activeElement == element is true

listbox2 should be focusable.
PASS document.activeElement == element is true

textarea2 should be focusable.
PASS document.activeElement == element is true

focusable2 should be focusable.
PASS document.activeElement == element is true

linkInHiddenCanvas should not be focusable.
PASS document.activeElement == previousFocusedElement is true

PASS successfullyParsed is true

TEST COMPLETE
